explicitly call storageBroker.delete(layerName) instead of truncate() on layer remove to take advantage of fast layer deletion
git-svn-id: http://svn.codehaus.org/geoserver/trunk@15565 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: c3962e322ac1bccb0389e75dbd45fb7e0b0e362f [formerly 06245434005d6029f2a9d2f3d2a82f51a204c933]
Former-commit-id: bf9aefc2860608857527dfbf1509f8cd966102e8

+++ src/web/gwc/src/main/java/org/geoserver/gwc/web/DiskQuotaConfigPanel.java
@@ -41,7 +41,7 @@ public class DiskQuotaConfigPanel extends Panel {
     @SuppressWarnings({ "rawtypes" })
     public DiskQuotaConfigPanel(final String id, final Form form,
             final IModel<DiskQuotaConfig> diskQuotaConfigModel, final IModel<GWC> gwcModel,
-            IModel<BigDecimal> configQuotaValueModel, IModel<StorageUnit> configQuotaUnitModel) {
+            IModel<Double> configQuotaValueModel, IModel<StorageUnit> configQuotaUnitModel) {
 
         super(id);
         this.gwcModel = gwcModel;
@@ -59,7 +59,7 @@ public class DiskQuotaConfigPanel extends Panel {
     }
 
     private void addGlobalQuotaConfig(final IModel<DiskQuotaConfig> diskQuotaModel,
-            IModel<BigDecimal> quotaValueModel, IModel<StorageUnit> unitModel) {
+            IModel<Double> quotaValueModel, IModel<StorageUnit> unitModel) {
 
         final IModel<Quota> globalQuotaModel = new LoadableDetachableModel<Quota>() {
             private static final long serialVersionUID = 1L;
@@ -84,7 +84,7 @@ public class DiskQuotaConfigPanel extends Panel {
                 "GWCSettingsPage.usedQuotaMessage", null, progressMessageParams);
         addGlobalQuotaStatusBar(globalQuotaModel, globalUsedQuotaModel, progressMessageModel);
 
-        TextField<BigDecimal> quotaValue = new TextField<BigDecimal>("globalQuota", quotaValueModel);
+        TextField<Double> quotaValue = new TextField<Double>("globalQuota", quotaValueModel);
         quotaValue.setRequired(true);
         add(quotaValue);
